requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
requests==2.31.0
gitpython==3.1.40